refactor: Computer Class 분리
- Computer Class를 랜덤 숫자를 생성하는 부분과 메세지를 출력하는 부분을 분리
- OutputView Class에 기존의 Computer Class에 있던 메시지를 출력하는 함수들 이동
- 기존의 Computer Class에 있던 출력을 담당하는 함수들의 이름을 함수의 역할을 더 잘 설명하는 이름으로 변경
